+++ tests/test_distro.py
@@ -11,14 +11,10 @@
 # CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 
-from __future__ import absolute_import
-from __future__ import print_function
-from __future__ import unicode_literals
-
-from os import path
-from os import listdir
+import os
 
 from commoncode.testcase import FileBasedTesting
+from commoncode.fileutils import resource_iter
 
 from container_inspector.distro import Distro
 from container_inspector.distro import parse_os_release
@@ -27,24 +23,24 @@ from utilities import check_expected
 
 
 class TestDistro(FileBasedTesting):
-    test_data_dir = path.join(path.dirname(__file__), 'data')
+    test_data_dir = os.path.join(os.path.dirname(__file__), 'data')
 
     def test_parse_os_release(self):
-        test_dir = self.get_test_loc('distro/chef-os-release/os-release')
-        expected_dir = self.get_test_loc('distro/chef-os-release/expected')
+        test_dir = self.get_test_loc('distro/os-release')
 
-        for os_release in listdir(test_dir):
-            test_file = path.join(test_dir, os_release)
-            expected = path.join(expected_dir, os_release + '-expected.json')
+        for test_file in resource_iter(test_dir, with_dirs=False):
+            if test_file.endswith('expected.json'):
+                continue
+            expected = test_file + '-expected.json'
             result = parse_os_release(test_file)
             check_expected(result, expected, regen=False)
 
     def test_distro_from_file(self):
-        test_dir = self.get_test_loc('distro/chef-os-release/os-release')
-        expected_dir = self.get_test_loc('distro/chef-os-release/expected-distro')
+        test_dir = self.get_test_loc('distro/os-release')
 
-        for os_release in listdir(test_dir):
-            test_file = path.join(test_dir, os_release)
-            expected = path.join(expected_dir, os_release + '-expected.json')
+        for test_file in resource_iter(test_dir, with_dirs=False):
+            if test_file.endswith('-expected.json'):
+                continue
+            expected = test_file + '-distro-expected.json'
             result = Distro.from_file(test_file).to_dict()
             check_expected(result, expected, regen=False)